+++ android/src/main/kotlin/com/example/flutter_video_compress/FlutterVideoCompressPlugin.kt
@@ -1,35 +1,26 @@
 package com.example.flutter_video_compress
 
-import android.annotation.SuppressLint
 import android.annotation.TargetApi
-import android.content.Context
 import android.graphics.Bitmap
-import android.media.MediaMetadataRetriever
 import android.media.MediaPlayer
 import android.media.ThumbnailUtils
 import android.net.Uri
 import android.os.Build
 import android.provider.MediaStore
-import android.text.TextUtils
-import android.util.Log
-import com.qiniu.pili.droid.shortvideo.PLShortVideoTranscoder
-import com.qiniu.pili.droid.shortvideo.PLVideoSaveListener
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
 import io.flutter.plugin.common.PluginRegistry.Registrar
+import nl.bravobit.ffmpeg.ExecuteBinaryResponseHandler
+import nl.bravobit.ffmpeg.FFmpeg
 import java.io.ByteArrayOutputStream
-import com.qiniu.pili.droid.shortvideo.PLErrorCode.ERROR_LOW_MEMORY
-import com.qiniu.pili.droid.shortvideo.PLErrorCode.ERROR_SRC_DST_SAME_FILE_PATH
-import com.qiniu.pili.droid.shortvideo.PLErrorCode.ERROR_NO_VIDEO_TRACK
-import kotlin.math.log
 import java.io.File
-import android.os.Environment
 
-class FlutterVideoCompressPlugin: MethodCallHandler {
-  private val channelName = "flutter_video_compress"
-  private val ENCODING_BITRATE_LEVEL_ARRAY = intArrayOf(320 * 480, 480 * 854, 544 * 960)
+
+class FlutterVideoCompressPlugin : MethodCallHandler {
+    private val channelName = "flutter_video_compress"
+    private var NOT_LOAD = true
 
     companion object {
         private lateinit var reg: Registrar
@@ -77,46 +68,33 @@ class FlutterVideoCompressPlugin: MethodCallHandler {
         result.success(mp.duration)
     }
 
-    @TargetApi(Build.VERSION_CODES.JELLY_BEAN_MR1)
     fun compressVideo(path: String, deleteOrigin: Boolean, result: Result) {
-        if (TextUtils.isEmpty(path)) {
-            return
+        val ffmpeg = FFmpeg.getInstance(reg.context())
+
+        if (!ffmpeg.isSupported) {
+            throw Exception("ffmpeg is supported")
+        }
+        val lastIndex = path.lastIndexOf("/")
+        val folder = path.substring(0, lastIndex) + "/flutter_video_compress"
+        val folderFile = File(folder);
+
+        if (!folderFile.exists()) {
+            File(folder).mkdirs()
         }
 
-        val lastIndex = path.lastIndexOf("/")
-        val newPath = path.substring(0, lastIndex) + "/flutter_video_compress" + path.substring(lastIndex)
+        val newPath = folder + path.substring(lastIndex)
 
-        val mShortVideoTransCoder = PLShortVideoTranscoder(reg.context(), path, newPath)
-        mShortVideoTransCoder.setMaxFrameRate(16)
-        val mediaRetriever = MediaMetadataRetriever()
-        mediaRetriever.setDataSource(path)
-        val height = mediaRetriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_VIDEO_HEIGHT)
-        val width = mediaRetriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_VIDEO_WIDTH)
-        val transCodingBitrateLevel = 1
-        val ratio = Integer.parseInt(width) / 720
+        val cmd = arrayOf("-i", path, "-vcodec", "libx264", "-crf", "28", "-acodec", "mp3", newPath)
 
-        mShortVideoTransCoder.transcode(720, Integer.parseInt(height) / ratio, getEncodingBitrateLevel(transCodingBitrateLevel), object : PLVideoSaveListener {
-            override fun onSaveVideoSuccess(s: String) {
+        ffmpeg.execute(cmd, object : ExecuteBinaryResponseHandler() {
+            override fun onFinish() {
+                result.success(newPath)
                 if (deleteOrigin) {
                     File(path).delete()
                 }
-                result.success(newPath)
-            }
-
-            override fun onSaveVideoFailed(errorCode: Int) {
-                result.error(channelName, errorCode.toString(), errorCode.toString())
-            }
-
-            override fun onSaveVideoCanceled() {
-            }
-
-            override fun onProgressUpdate(percentage: Float) {
 
             }
         })
-    }
 
-    private fun getEncodingBitrateLevel(position: Int): Int {
-        return ENCODING_BITRATE_LEVEL_ARRAY[position]
     }
 }