+++ example/lib/main.dart
@@ -1,4 +1,3 @@
-
 import 'dart:io';
 import 'dart:typed_data';
 
@@ -31,18 +30,33 @@ class _MyAppState extends State<MyApp> {
     }
   }
 
-  @override
-  Widget build(BuildContext context) {
+  List<Widget> _builColumnChildren() {
+    // dart 2.3 before
     final _list = <Widget>[
-      FlatButton(child: Text('take video'), onPressed: _videoPicker),
+      FlatButton(child: Text('take video'), onPressed: _videoPicker)
     ];
     if (_image != null) {
-      _list.add(Flexible(child: Center(child: Image.memory(_image))));
+      _list.add(Flexible(child: Image.memory(_image)));
     }
+    return _list;
+
+    // dart 2.3
+    // final _list =  [
+    //   FlatButton(child: Text('take video'), onPressed: _videoPicker),
+    //   if(_image != null) Flexible(child: Image.memory(_image))
+    // ];
+    // return _list;
+  }
+
+  @override
+  Widget build(BuildContext context) {
     return MaterialApp(
       home: Scaffold(
         appBar: AppBar(title: const Text('Plugin example app')),
-        body: Column(children: _list),
+        body: Column(
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          children: _builColumnChildren(),
+        ),
       ),
     );
   }