+++ test/flutter_video_compress_test.dart
@@ -1,9 +1,9 @@
-import 'package:flutter/services.dart';
+//import 'package:flutter/services.dart';
 // import 'package:flutter_test/flutter_test.dart';
 // import 'package:flutter_video_compress/flutter_video_compress.dart';
 
 void main() {
-  const MethodChannel channel = MethodChannel('flutter_video_compress');
+//  const MethodChannel channel = MethodChannel('flutter_video_compress');
 
   // setUp(() {
   //   channel.setMockMethodCallHandler((MethodCall methodCall) async {